feature: generate thumbnails and display all stories on the homepage
Uses the `generateStories.js` script and Dall-E 3 to generate thumbnails
based on the story's description. The image is stored in
`/public/stories/images`.

Uses the `generateStories.js` script to add new entries to a JSON
metadata file that keeps an array of the metadata for all stories.
This is used to render a list of stories on the home page.

Closes #2, #3